[ARA] Adding a simple region sequence wrapper class - this is still a WIP.
In my impementation I'm using playback regions to manage the lifetime of region sequences - each playback region increments a count stored by the RegionSequenceWrapper, and when this goes to zero the region sequence is destroyed.
The problem is that I'm using the Track that the playback region's Clip belongs to as a key between playback regions and region sequences, and it seems like Tracks are destroyed prior to ARA playback regions being destroyed. Thus when the playback region tries to decrement its region sequence's reference count in its destructor it can't find a region sequence because its track no longer exists.
I'm not familiar enough with the tracktion_engine code to know where the best place for region sequence lifetime management should be,  but I think my implementation will have to be adjusted. Note that nothing is leaked because we're using std::unique_ptr to manage our RegionSequenceWrapper instances, but the situation is shaky nonetheless.