Fix documentation python crashing if module header is not found, instead gracefully indicate the the module header (i.e. the juce dependency) is skipped.
Also add a convenient debug launch.json vscode python config file to allow debugging python script(s). In order to use vscode to build/debug python just install in vscode the "Python" and "Python for VSCode" extensions then open a python file.

In future, we should be able to generate the modules grouping directly from source and therefore not duplicate the full source code to the build folder, this seems unnecessary and slows down significantly the documentation process.
The usual approach for adding modules groups is to ask each class to add itself to a group (using the @addgroup statement) so that doxygen will automatically discovers them eliminating the need for the script. while preserving complete genericity to that process.

For now keep using the script and focus of the documentation content.