Update tracktion_Phaser.cpp
When `depth` is set to 0 or very close to 0 (probably CPU dependant) the assignment of swpFactor to be the value of sweepFactor at PhaserPlugin::applyToBuffer():24 causes swpFactor to always be 1. Consequently when `swp` is multiply assigned by `swpFactor` with a value of 1, the value of `swp` is always greater than the value of `maxSweep` for this `depth`, and so `swpFactor` is always assigned the value of `sweepDown`. The impact is to have the phaser bottom out and get stuck at  -2.77556e-16 on an x86 CPU.

At a `depth` of 0 range's value is 1. Given a range of 1 the value of `maxSweep` value is identical to `minSweep`. Given these details as the value of `depth` approaches 0, and depending on the accuracy of the CPU's implementation of denormal numbers, the Phaser rapidly approaches a point where it no longer works as expected.

My solution is to remove PhaserPlugin::appyToBuffer():24 `swpFactor = sweepFactor;`. I'm not sure why the value of `swpFactor` is filled again before it's read, but removing this makes the ternary operator at PhaserPlugin::appyToBuffer():15 successful at switching the direction of the factor as it crosses between upwards and downwards phase sweeping. This change should not change any of the aesthetic qualities of the Phaser.